ADDED: main function, located in main.Reporter, and user interface

+++ ENGR3720U/Part A/Problem 2/src/fitnessFunctions/F5.java
@@ -2,11 +2,11 @@ package fitnessFunctions;
 
 public class F5 implements FitnessFunction {
     @Override
-    public Double apply(final Double[] parameters) {
-        Double sum = (double) (10 * parameters.length);
+    public Double apply(final double[] parameters) {
+        double sum = 10 * parameters.length;
 
         for (int i = 1; i <= parameters.length - 1; i++) {
-            final Double param = parameters[i - 1];
+            final double param = parameters[i - 1];
 
             sum += Math.pow(param, 2) - 10 * Math.cos(2 * Math.PI * param);
         }
